+++ packages/loaders/tests/zarr-lib.spec.js
@@ -1,4 +1,4 @@
-import { test } from 'tape';
+import { describe, test, expect } from 'vitest';
 import { FileSystemStore } from './common';
 import { loadMultiscales } from '../src/zarr/lib/utils';
 import { getIndexer } from '../src/zarr/lib/indexer';
@@ -8,37 +8,31 @@ import * as url from 'url';
 
 const __dirname = url.fileURLToPath(path.dirname(import.meta.url));
 const FIXTURE = path.resolve(__dirname, './fixtures/bioformats-zarr');
-test('Loads zarr-multiscales', async t => {
-  t.plan(1);
-  try {
-    const store = new FileSystemStore(`${FIXTURE}/data.zarr`);
-    const { data } = await loadMultiscales(store, '0');
-    t.equal(data.length, 2, 'Should have two multiscale images.');
-  } catch (e) {
-    t.fail(e);
-  }
+
+test('loadMultiscales', async () => {
+  const store = new FileSystemStore(`${FIXTURE}/data.zarr`);
+  const { data } = await loadMultiscales(store, '0');
+  expect(data.length).toBe(2);
 });
 
-test('Indexer creation and usage.', async t => {
-  t.plan(4);
-  const labels = ['a', 'b', 'y', 'x'];
-  const indexer = getIndexer(labels);
-  t.deepEqual(
-    indexer({ a: 10, b: 20 }),
-    [10, 20, 0, 0],
-    'should allow named indexing.'
-  );
-  t.deepEqual(
-    indexer([10, 20, 0, 0]),
-    [10, 20, 0, 0],
-    'allows array like indexing.'
-  );
-  t.throws(
-    () => indexer({ c: 0, b: 0 }),
-    'should throw with invalid dim name.'
-  );
-  t.throws(
-    () => getIndexer(['a', 'b', 'c', 'b', 'y', 'x']),
-    'no duplicated labels names.'
-  );
+describe('getIndexer', async () => {
+  test.each([
+    [{ a: 10, b: 20 }, [10, 20, 0, 0]],
+    [
+      [10, 20, 0, 0],
+      [10, 20, 0, 0]
+    ]
+  ])(`indexer(%j)`, (input, expected) => {
+    const indexer = getIndexer(['a', 'b', 'y', 'x']);
+    expect(indexer(input)).toStrictEqual(expected);
+  });
+
+  test('throws with invalid dim name', () => {
+    const indexer = getIndexer(['a', 'b', 'y', 'x']);
+    expect(() => indexer({ c: 0, b: 0 })).toThrowError();
+  });
+
+  test('throws with duplicate dim name', () => {
+    expect(() => getIndexer(['a', 'b', 'c', 'b', 'y', 'x'])).toThrowError();
+  });
 });