+++ sites/avivator/main.js
@@ -1,0 +1,3 @@
+import { loadOmeTiff } from '@hms-dbmi/viv';
+
+await loadOmeTiff('http://localhost:5173/@data/multifile.companion.ome');