+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 ### Changed
+- Bump `gl` to v6 and move to Node.js 18 in CI.
 
 ## 0.13.8